ddk-node: restructure project
