+++ ddk-node/src/seed.rs
